HADOOP-10437. Fix the javac warnings in the conf and the util package.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/common/trunk@1582015 13f79535-47bb-0310-9956-ffa450edef68

+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/util/Progress.java
@@ -64,7 +64,7 @@ public class Progress {
   public synchronized Progress addPhase() {
     Progress phase = addNewPhase();
     // set equal weightage for all phases
-    progressPerPhase = 1.0f / (float)phases.size();
+    progressPerPhase = 1.0f / phases.size();
     fixedWeightageForAllPhases = true;
     return phase;
   }
@@ -110,7 +110,7 @@ public class Progress {
       addNewPhase();
     }
     // set equal weightage for all phases
-    progressPerPhase = 1.0f / (float)phases.size();
+    progressPerPhase = 1.0f / phases.size();
     fixedWeightageForAllPhases = true;
   }
